preserve hyphens in FORM and LEMMA columns
Preserve FORMs and LEMMAs which are hyphens and en-dashes, instead of
interpreting these as mangled null values which need to be normalised to
"_".

Originally reported as a bug in Arborator; annotators were surprised and
distressed when their hyphens and en-dashes were transformed to
underscores on save. Really aggressive normalisation probably belongs
closer to the frontend, rather than in the specification library.

Reported-By: Yifat Ben Moshe <yifat1811@gmail.com>